Make Metrics a separate crate
