+++ Cargo.toml
@@ -19,6 +19,7 @@ members = [
   "oak_functions/examples/mobilenet/module",
   "oak_functions/examples/weather_lookup/module",
   "oak_functions/experimental/tf_inference",
+  "oak_functions/experimental/metrics",
   "oak_functions/load_test",
   "oak_functions/loader",
   "oak_functions/location_utils",
