+++ oak_functions/loader/Cargo.toml
@@ -11,7 +11,7 @@ license = "Apache-2.0"
 oak-unsafe = ["oak-tf", "oak-metrics"]
 default = []
 oak-tf = ["oak_functions_tf_inference"]
-oak-metrics = []
+oak-metrics = ["oak_functions_metrics"]
 
 [dependencies]
 anyhow = "*"
@@ -37,6 +37,7 @@ humantime-serde = "*"
 log = { version = "*", features = ["max_level_off", "release_max_level_off"] }
 oak_functions_abi = { path = "../abi" }
 oak_functions_tf_inference = { optional = true, path = "../experimental/tf_inference" }
+oak_functions_metrics = { optional = true, path = "../experimental/metrics" }
 oak_logger = { path = "../logger" }
 oak_remote_attestation = { path = "../../remote_attestation/rust/" }
 prost = "*"
