+++ oak_functions/loader/src/tf.rs
@@ -32,7 +32,7 @@ use wasmi::ValueType;
 /// Host function name for invoking TensorFlow model inference.
 const TF_ABI_FUNCTION_NAME: &str = "tf_model_infer";
 
-// TODO(#2576): Move extension implementation to `tf_inference` crate once the Extension-realted
+// TODO(#2576): Move extension implementation to `tf_inference` crate once the Extension-related
 // structs are in a separate crate.
 impl OakApiNativeExtension for TensorFlowModel<Logger> {
     fn invoke(
